+++ apps/web/app/[locale]/not-found.tsx
@@ -1,17 +1,5 @@
-'use client';
+import { notFound } from 'next/navigation';
 
-import NotFound from '@components/pages/404';
-import { AuthLayout } from 'lib/layout';
-import { useTranslations } from 'next-intl';
-
-const NotFoundPage = () => {
-	const t = useTranslations();
-
-	return (
-		<AuthLayout title={t('pages.page404.HEADING_TITLE')} isAuthPage={false}>
-			<NotFound />
-		</AuthLayout>
-	);
-};
-
-export default NotFoundPage;
+export default function CatchAllPage() {
+	notFound();
+}
