+++ apps/web/app/not-found.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import NotFound from '@components/pages/404';
+
+const NotFoundPage = () => {
+	return (
+		<html>
+			<body>
+				<NotFound />
+			</body>
+		</html>
+	);
+};
+
+export default NotFoundPage;
